+++ src/main/java/org/mybatis/dynamic/sql/SqlBuilder.java
@@ -1,5 +1,5 @@
 /*
- *    Copyright 2016-2023 the original author or authors.
+ *    Copyright 2016-2024 the original author or authors.
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -490,6 +490,10 @@ public interface SqlBuilder {
 
     static StringConstant stringConstant(String constant) {
         return StringConstant.of(constant);
+    }
+
+    static <T> BoundValue<T> value(T value) {
+        return BoundValue.of(value);
     }
 
     // functions